pyproject.toml
[project]
name = "finance-advisor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.0.1",
    "langgraph>=1.0.10",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "pandas>=2.3.3",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "streamlit>=1.54.0",
    "supabase>=2.28.0",
    "tavily-python>=0.7.22",
    "uvicorn>=0.41.0",
]
